Refactors YoutubeResources to VideoResources to make space for Youtube functionality and clarity
