+++ harvester/files/fixtures/test-file-document.json
@@ -162,7 +162,7 @@
                         "is_video"
                     ],
                     "resources": [
-                        "files.YoutubeThumbnailResource"
+                        "files.VideoThumbnailResource"
                     ],
                     "depends_on": []
                 }
