+++ harvester/files/models/__init__.py
@@ -6,7 +6,7 @@ from files.models.datatypes.file import FileDocument, Overwrite
 
 from files.models.resources.metadata import HttpTikaResource, CheckURLResource
 from files.models.resources.pdf_thumbnail import PdfThumbnailResource
-from files.models.resources.youtube_thumbnail import YoutubeThumbnailResource
-from files.models.resources.youtube_api import YoutubeAPIResource
-from files.models.resources.youtube_transcripts import YoutubeTranscriptsResource
+from files.models.resources.video_thumbnail import VideoThumbnailResource
+from files.models.resources.youtube import YoutubeAPIResource
+from files.models.resources.video_transcripts import VideoTranscriptsResource
 from files.models.resources.image_thumbnail import ImageThumbnailResource
