+++ harvester/files/tests/resources/test_youtube_thumbnail.py
@@ -1,17 +1,17 @@
 from django.test import TestCase
 
-from files.models import YoutubeThumbnailResource
+from files.models import VideoThumbnailResource
 
 
 class TestYoutubeThumbnailResource(TestCase):
 
     def test_get_preview_filename(self):
-        jpeg_filename = YoutubeThumbnailResource.get_preview_filename(
+        jpeg_filename = VideoThumbnailResource.get_preview_filename(
             "https://i.ytimg.com/vi/cZEclkrS_xY/hqdefault.jpg?"
             "sqp=-oaymwEcCNACELwBSFXyq4qpAw4IARUAAIhCGAFwAcABBg==&rs=AOn4CLBzU2IL91e3jXFPvl4LYENzoyvoeg"
         )
         self.assertEqual(jpeg_filename, "cZEclkrS_xY.jpg")
-        webp_filename = YoutubeThumbnailResource.get_preview_filename(
+        webp_filename = VideoThumbnailResource.get_preview_filename(
             "https://i.ytimg.com/vi_webp/t6tUThsvE4M/maxresdefault.webp"
         )
         self.assertEqual(webp_filename, "t6tUThsvE4M.webp")
